requirements.txt
streamlit
pandas
numpy
plotly
google-generativeai
python-dotenv
torch
scikit-learn
statsmodels
